Streamlit/requirements.txt
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
plotly==5.18.0
streamlit==1.29.0
